+++ include/unifex/let_with_stop_source.hpp
